Refactor native asset handling for datablocks and mediaforce
